// src/components/Forms/Sumup/tipo-demanda/interfaces.ts
export interface Autor {
  id: string;
  nome: string;
  documento: string;
  email: string;
  telefone: string;
}

export interface Reu {
  id: string;
  nome: string;
  documento: string;
}
// ...
export interface DadosDemandaProcon {
  pk?: string;
  usuario_atuando?: string;
  created_by?: string;
  cliente?: string;
  data_carimbo?: string;
  due_date?: string;
  arquivos?: ArquivoDemanda[];
  tipo_demanda?: string;
  perfil_demanda?: string;
  status?: string;
  status_demanda?: string;
  identificacao?: string;
  
  campos_auditados?: any;

  numero_reserva?: string;
  data_defesa?: string;
  data_audiencia?: string;
  tipificacao?: string;
  uf?: string;
  cidade?: string;
  objeto?: string;
  data_distribuicao?: string;
  data_reclamacao?: string;
  resumo_processo?: string;
  tipo_processo?: string;
  origem_reclamacao?: string;
  causa_raiz?: string;
  sub_causa_raiz?: string;
  lista_autores?: Autor[];
  lista_reus?: Reu[];
  observacao?: {
    mensagem: string;
    criada_em: string;
    criada_por: string;
  }[];
}
// ...
export interface FormFile {
  id: string;
  info: {
    name: string;
    size: number;
    sizeFormatted: string;
    url: string; 
    unit: Unit;
    s3Key: string;
    s3Bucket: string;
  };
  status: {
    success: boolean | undefined;
    progress: number;
    message: string | undefined;
    isDeleting?: boolean;
  };
  allow: {
    delete: boolean;
    download: boolean;
    retryUpload: boolean;
    link: boolean;
  };
  dropzoneFile: File | FileRejection | undefined;
}
// ...
export type TipoJustificativaInfo = 'enviar-excecao-cliente' | 
'enviar-esteira-cliente' | 'enviar-esteira-oito' | 'enviar-auditoria-oito' |
'enviar-email' | 'substituir-endereco-email' | 'inativar-demanda' |
'enviar-excecao-oito' | 'reprovar-demanda' | '';

export interface JustificativaInfo {
  tipo: TipoJustificativaInfo;
  title: Partial<{
    text: string;
  }>,
  justificativa: Partial<{
    isValid: boolean;
    value: string;
  }>
}